+++ src/mcpmodel/__init__.py
@@ -1,4 +1,3 @@
 """MCPModel research utilities."""
 
 __version__ = "0.1.0.dev0"
-